// framework/magento/schema.d.ts
export type CategoryTree = {
  available_sort_by: Array<Scalars["String"]>;
  breadcrumbs: Array<Breadcrumb>;
  canonical_url: Scalars["String"];
  children: Array<CategoryTree>;
  children_count: Scalars["String"];
  cms_block: CmsBlock;
  custom_layout_update_file: Scalars["String"];
  default_sort_by: Scalars["String"];
  description: Scalars["String"];
  display_mode: Scalars["String"];
  filter_price_range: Scalars["Float"];
  image: Scalars["String"];
  include_in_menu: Scalars["Int"];
  is_anchor: Scalars["Int"];
  landing_page: Scalars["Int"];
  level: Scalars["Int"];
  meta_description: Scalars["String"];
  meta_keywords: Scalars["String"];
  meta_title: Scalars["String"];
  name: Scalars["String"];
  path: Scalars["String"];
  path_in_store: Scalars["String"];
  position: Scalars["Int"];
  product_count: Scalars["Int"];
  products: CategoryProducts;
  relative_url: Scalars["String"];
  titlecat: Scalars["String"];
  type: UrlRewriteEntityTypeEnum;
  uid: Scalars["String"];
  url_key: Scalars["String"];
  url_path: Scalars["String"];
  url_suffix: Scalars["String"];
};

export type CmsBlock = {
  content: Scalars["String"];
  identifier: Scalars["String"];
  title: Scalars["String"];
};

export enum UrlRewriteEntityTypeEnum {
  CMS_PAGE = "CMS_PAGE",
  PRODUCT = "PRODUCT",
  CATEGORY = "CATEGORY",
}

export type ProductAttributeSortInput = {
  maxguests: SortEnum;
  name: SortEnum;
  position: SortEnum;
  price: SortEnum;
  relevance: SortEnum;
};

export enum SortEnum {
  ASC = "ASC",
  DESC = "DESC",
}

export type CategoryProducts = {
  items: Array<ProductItem>;
  page_info: SearchResultPageInfo;
  total_count: Scalars["Int"];
};

export type MultiSelectAttributes<T> = {
  MultiSelectAttributes: MultiSelectAttributes;
  [key: T]: T;
};

export type ComplexTextValue = {
  html: Scalars["String"];
};
// ...
export type Money = {
  currency: CurrenyEnum;
  value: Scalars["Float"];
};
// ...
export type ProductItem = {
  additional_information: Scalars["String"];
  builder: Scalars["String"];
  canonical_url: Scalars["String"];
  categories: Array<CategoryTree>;
  country_of_manufacture: Scalars["String"];
  crosssell_products: Array<ProductItem[]>;
  customattributes: MultiSelectAttributes;
  description: ComplexTextValue;
  exterior_designer: Scalars["String"];
  gift_message_available: Scalars["String"];
  image: ProductImage;
  media_gallery: ProductImage[];
  meta_description: Scalars["String"];
  meta_keyword: Scalars["String"];
  meta_title: Scalars["String"];
  name: Scalars["String"];
  new_from_date: Scalars["String"];
  new_to_date: Scalars["String"];
  only_x_left_in_stock: Float;
  options_container: Scalars["String"];
  price_range: PriceRange;
  price_tiers: Array<TierPrice>;
  product_links: Array<ProductLinks>;
  productattribute: Scalars["String"];
  related_products: ProductItem[];
  short_description: ComplexTextValue;
  sku: Scalars["String"];
  small_image: ProductImage;
  special_price: Scalars["Float"];
  special_to_date: Scalars["String"];
  stock_status: ProductStockStatus;
  swatch_image: Scalars["String"];
  thumbnail: ProductImage;
  uid: Scalars["ID"];
  upgrades: Scalars["String"];
  upsell_products: ProductItem[];
  url_key: Scalars["String"];
  url_rewrites: Array<UrlRewrite>;
  url_suffix: Scalars["String"];
  videoid: Scalars["String"];
  year: Scalars["String"];
};
// ...
export type SearchResultPageInfo = {
  current_page: Scalars["Int"];
  page_size: Scalars["Int"];
  total_pages: Scalars["Int"];
};

export type Breadcrumb = {
  category_id: Scalars["Int"];
  category_level: Scalars["Int"];
  category_name: Scalars["String"];
  category_uid: Scalars["String"];
  category_url_key: Scalars["String"];
  category_url_path: Scalars["String"];
};
// ...
export type Attribute = {
  attribute_code: Scalars["String"];
  attribute_options: Array<AttributeOption>;
  attribute_type: Scalars["String"];
  entity_type: Scalars["String"];
  input_type: Scalars["String"];
};

export type customAttributeMetadata = {
  items: Array<Attribute>;
};

//shopify
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An ISO-8601 encoded UTC date time string. Example value: `"2019-07-03T20:47:55Z"`. */
  DateTime: any;
  /** A signed decimal number, which supports arbitrary precision and is serialized as a string. Example value: `"29.99"`. */
  Decimal: any;
  /** A string containing HTML code. Example value: `"<p>Grey cotton knit sweater.</p>"`. */
  HTML: any;
  /** A monetary value string. Example value: `"100.57"`. */
  Money: any;
  /**
   * An RFC 3986 and RFC 3987 compliant URI string.
   *
   * Example value: `"https://johns-apparel.myshopify.com"`.
   *
   */
  URL: any;
};
// ...
export type AppliedGiftCard = Node & {
  __typename?: "AppliedGiftCard";
  /**
   * The amount that was taken from the gift card by applying it.
   * @deprecated Use `amountUsedV2` instead
   */
  amountUsed: Scalars["Money"];
  /** The amount that was taken from the gift card by applying it. */
  amountUsedV2: MoneyV2;
  /**
   * The amount left on the gift card.
   * @deprecated Use `balanceV2` instead
   */
  balance: Scalars["Money"];
  /** The amount left on the gift card. */
  balanceV2: MoneyV2;
  /** Globally unique identifier. */
  id: Scalars["ID"];
  /** The last characters of the gift card. */
  lastCharacters: Scalars["String"];
  /** The amount that was applied to the checkout in its currency. */
  presentmentAmountUsed: MoneyV2;
};
// ...
export type AvailableShippingRates = {
  __typename?: "AvailableShippingRates";
  /**
   * Whether or not the shipping rates are ready.
   * The `shippingRates` field is `null` when this value is `false`.
   * This field should be polled until its value becomes `true`.
   */
  ready: Scalars["Boolean"];
  /** The fetched shipping rates. `null` until the `ready` field is `true`. */
  shippingRates?: Maybe<Array<ShippingRate>>;
};
// ...
export type Checkout = Node & {
  __typename?: "Checkout";
  /** The gift cards used on the checkout. */
  appliedGiftCards: Array<AppliedGiftCard>;
  /**
   * The available shipping rates for this Checkout.
   * Should only be used when checkout `requiresShipping` is `true` and
   * the shipping address is valid.
   */
  availableShippingRates?: Maybe<AvailableShippingRates>;
  /** The date and time when the checkout was completed. */
  completedAt?: Maybe<Scalars["DateTime"]>;
  /** The date and time when the checkout was created. */
  createdAt: Scalars["DateTime"];
  /** The currency code for the Checkout. */
  currencyCode: CurrencyCode;
  /** A list of extra information that is added to the checkout. */
  customAttributes: Array<Attribute>;
  /**
   * The customer associated with the checkout.
   * @deprecated This field will always return null. If you have an authentication token for the customer, you can use the `customer` field on the query root to retrieve it.
   */
  customer?: Maybe<Customer>;
  /** Discounts that have been applied on the checkout. */
  discountApplications: DiscountApplicationConnection;
  /** The email attached to this checkout. */
  email?: Maybe<Scalars["String"]>;
  /** Globally unique identifier. */
  id: Scalars["ID"];
  /** A list of line item objects, each one containing information about an item in the checkout. */
  lineItems: CheckoutLineItemConnection;
  /** The sum of all the prices of all the items in the checkout. Duties, taxes, shipping and discounts excluded. */
  lineItemsSubtotalPrice: MoneyV2;
  /** The note associated with the checkout. */
  note?: Maybe<Scalars["String"]>;
  /** The resulting order from a paid checkout. */
  order?: Maybe<Order>;
  /** The Order Status Page for this Checkout, null when checkout is not completed. */
  orderStatusUrl?: Maybe<Scalars["URL"]>;
  /**
   * The amount left to be paid. This is equal to the cost of the line items, taxes and shipping minus discounts and gift cards.
   * @deprecated Use `paymentDueV2` instead
   */
  paymentDue: Scalars["Money"];
  /** The amount left to be paid. This is equal to the cost of the line items, duties, taxes and shipping minus discounts and gift cards. */
  paymentDueV2: MoneyV2;
  /**
   * Whether or not the Checkout is ready and can be completed. Checkouts may
   * have asynchronous operations that can take time to finish. If you want
   * to complete a checkout or ensure all the fields are populated and up to
   * date, polling is required until the value is true.
   */
  ready: Scalars["Boolean"];
  /** States whether or not the fulfillment requires shipping. */
  requiresShipping: Scalars["Boolean"];
  /** The shipping address to where the line items will be shipped. */
  shippingAddress?: Maybe<MailingAddress>;
  /** The discounts that have been allocated onto the shipping line by discount applications. */
  shippingDiscountAllocations: Array<DiscountAllocation>;
  /** Once a shipping rate is selected by the customer it is transitioned to a `shipping_line` object. */
  shippingLine?: Maybe<ShippingRate>;
  /**
   * Price of the checkout before shipping and taxes.
   * @deprecated Use `subtotalPriceV2` instead
   */
  subtotalPrice: Scalars["Money"];
  /** Price of the checkout before duties, shipping and taxes. */
  subtotalPriceV2: MoneyV2;
  /** Specifies if the Checkout is tax exempt. */
  taxExempt: Scalars["Boolean"];
  /** Specifies if taxes are included in the line item and shipping line prices. */
  taxesIncluded: Scalars["Boolean"];
  /**
   * The sum of all the prices of all the items in the checkout, taxes and discounts included.
   * @deprecated Use `totalPriceV2` instead
   */
  totalPrice: Scalars["Money"];
  /** The sum of all the prices of all the items in the checkout, duties, taxes and discounts included. */
  totalPriceV2: MoneyV2;
  /**
   * The sum of all the taxes applied to the line items and shipping lines in the checkout.
   * @deprecated Use `totalTaxV2` instead
   */
  totalTax: Scalars["Money"];
  /** The sum of all the taxes applied to the line items and shipping lines in the checkout. */
  totalTaxV2: MoneyV2;
  /** The date and time when the checkout was last updated. */
  updatedAt: Scalars["DateTime"];
  /** The url pointing to the checkout accessible from the web. */
  webUrl: Scalars["URL"];
};
// ...
export type CheckoutLineItemConnection = {
  __typename?: "CheckoutLineItemConnection";
  /** A list of edges. */
  edges: Array<CheckoutLineItemEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type DiscountAllocation = {
  __typename?: "DiscountAllocation";
  /** Amount of discount allocated. */
  allocatedAmount: MoneyV2;
  /** The discount this allocated amount originated from. */
  discountApplication: DiscountApplication;
};
// ...
export type DiscountApplicationConnection = {
  __typename?: "DiscountApplicationConnection";
  /** A list of edges. */
  edges: Array<DiscountApplicationEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};
// ...
export type MoneyV2 = {
  __typename?: "MoneyV2";
  /** Decimal money amount. */
  amount: Scalars["Decimal"];
  /** Currency of the money. */
  currencyCode: CurrencyCode;
};
// ...
export type Node = {
  /** Globally unique identifier. */
  id: Scalars["ID"];
};
// ...
export type PageInfo = {
  __typename?: "PageInfo";
  /** Indicates if there are more pages to fetch. */
  hasNextPage: Scalars["Boolean"];
  /** Indicates if there are any pages prior to the current page. */
  hasPreviousPage: Scalars["Boolean"];
};
// ...
export type ProductOption = Node & {
  __typename?: "ProductOption";
  /** Globally unique identifier. */
  id: Scalars["ID"];
  /** The product option’s name. */
  name: Scalars["String"];
  /** The corresponding value to the product option name. */
  values: Array<Scalars["String"]>;
};
// ...
export type ProductVariantConnection = {
  __typename?: "ProductVariantConnection";
  /** A list of edges. */
  edges: Array<ProductVariantEdge>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
};

/** An auto-generated type which holds one ProductVariant and a cursor during pagination. */
export type ProductVariantEdge = {
  __typename?: "ProductVariantEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of ProductVariantEdge. */
  node: ProductVariant;
};
// ...
export type SelectedOption = {
  __typename?: "SelectedOption";
  /** The product option’s name. */
  name: Scalars["String"];
  /** The product option’s value. */
  value: Scalars["String"];
};